+++ Sources/Swift/Integrations/WatchdogTerminations/Processors/SentryWatchdogTerminationBaseProcessor.swift
@@ -1,0 +1,45 @@
+@_implementationOnly import _SentryPrivate
+import Foundation
+
+class SentryWatchdogTerminationBaseProcessor<T>: NSObject {
+    
+    private let dispatchQueueWrapper: SentryDispatchQueueWrapper
+    private let store: SentryScopeBasePersistentStore
+    private let dataTypeName: String
+    
+    init(
+        withDispatchQueueWrapper dispatchQueueWrapper: SentryDispatchQueueWrapper,
+        store: SentryScopeBasePersistentStore,
+        dataTypeName: String
+    ) {
+        self.dispatchQueueWrapper = dispatchQueueWrapper
+        self.store = store
+        self.dataTypeName = dataTypeName
+        
+        super.init()
+        
+        clear()
+    }
+    
+    func setData(_ data: T?, writeAction: @escaping (T) -> Void) {
+        SentrySDKLog.debug("Setting \(dataTypeName) in background queue: \(String(describing: data))")
+        let name = dataTypeName
+        dispatchQueueWrapper.dispatchAsync { [weak self] in
+            guard let strongSelf = self else {
+                SentrySDKLog.debug("Can not set \(name), reason: reference to \(name) processor is nil")
+                return
+            }
+            guard let data = data else {
+                SentrySDKLog.debug("\(strongSelf.dataTypeName) is nil, deleting active file.")
+                strongSelf.store.deleteStateOnDisk()
+                return
+            }
+            writeAction(data)
+        }
+    }
+    
+    public func clear() {
+        SentrySDKLog.debug("Deleting \(dataTypeName) file in persistent store")
+        store.deleteStateOnDisk()
+    }
+}